Added bill calculation

+++ app/models/item_model.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 from uuid import UUID
-
+from decimal import Decimal
 from sqlalchemy import String
 from sqlmodel import Field, Column, Relationship
 
@@ -12,13 +12,19 @@ if TYPE_CHECKING:
 
 
 class ItemBase(BaseModel):
-    name: str = Field(sa_column=Column(String(50), index=True, unique=True, nullable=False))
-    price: float = Field(default=0, nullable=False)
+    name: str = Field(
+        sa_column=Column(String(50), index=True, unique=True, nullable=False)
+    )
+    price: Decimal = Field(default=0, nullable=False)
     assigned_user_uuid: UUID = Field(default=None, foreign_key="user.uuid")
     bill_uuid: UUID = Field(default=None, foreign_key="bill.uuid")
     is_paid: bool = Field(default=False, nullable=False)
 
 
 class Item(BaseUUIDModel, ItemBase, table=True):
-    assigned_user: "User" = Relationship(back_populates="items", sa_relationship_kwargs={"lazy": "selectin"})
-    bill: "Bill" = Relationship(back_populates="items", sa_relationship_kwargs={"lazy": "selectin"})
+    assigned_user: "User" = Relationship(
+        back_populates="items", sa_relationship_kwargs={"lazy": "selectin"}
+    )
+    bill: "Bill" = Relationship(
+        back_populates="items", sa_relationship_kwargs={"lazy": "selectin"}
+    )
